Reduce allocationId to 24 bits

+++ contracts/interfaces/ICover.sol
@@ -79,7 +79,7 @@ struct PoolAllocation {
   uint40 poolId;
   uint96 coverAmountInNXM;
   uint96 premiumInNXM;
-  uint allocationId;
+  uint24 allocationId;
 }
 
 struct CoverData {
